nm applet is basically functional
This makes the applet basically functional: it loads, imports the
plugin, shows the connection state in compact representation and list
the available connections in the popup. Changing connection works.

Traffic monitor is disabled since we don't have a suitable Plotter
component yet, this is work in progress.

For now, I've disabled ModemManager support, since there are some
linking problems, and the include and ifdef situation is not completely
sorted.