Stylistic changes to the ConnectionItem
- bold only when connection is active
The font weight and italic is used to indicate connection state.
Painting all items with a Uuid bold defeats the purpose, since it's now
not immediately clean when a connection is active from the boldness of
hte font.

- opacity change for the subtext: this change is in line with style
  recommendations from the artists, subitems opacity is used as .6
throughout plasma, use that instead of color-string-hackery.

- move visible property into buttonRow
All items have this change set, they might inherit it just as well from
their common parent.