Change ordering of buttons to be more predictable
Switch info and edit buttons of the buttonRow in the delegate item.
Connections usually have info, but not necessarily edit. This change
results in the info button always showing up in the same place in the
items, making for a more uniform look and more predictable positioning
of actions, supporting the user's spatial memory.

Conflicts:
	applet/declarative/contents/ui/ConnectionItem.qml