Remove redirect-gateway tag from exported/imported config file
This is not used even in nm-connection-editor. Also if we are checking for this tag
which will never be present in exported OpenVPN connection from nm-connection-editor
we always set default-route in IPv4 config as true, which is not usually what users
want. I was talking with Dan Williams yesterday and NM-openvpn relies on autodetection
of the server user is connecting to. Also he plan to add functionality to this plugin
to set never-default automatically from this plugin.

BUG:340801