Move the view to the top while a connection is being activated
Use a slightly different approach to see whether it works. Previously
we used positionViewAtIndex(index, ListView.End) which made the connection
visible and moved the view so the connection was at the bottom, but there
were some glitches at that time. Let's try to move the view at the
beginning, which should show the connection which is being activated in
case that user has less than five active connections.

BUG:342801