Improve connection deactivated/activated messages when suspending
Don't emit connection deactivated signal when we're about to suspend.
Also don't emit the activated notification when waking up when we're connecting
to the same connection we've been connected to prior to suspending. (there's a lot
of connections in that sencence).

If, after waking up, there's no active or activating connection (auto reconnecting
to another WiFi) it will notify the user that he's no longer connected to a network.

REVIEW: 126748