openvpn: rename comp-lzo setting "No by default" in the UI
To express a "comp-lzo no" setting in nm-openvpn, it must be stored as
"comp-lzo=no-by-default". That is to preserve backward compatiblity,
where older versions would wrongly set "comp-lzo=no" to mean LZO
compression disabled [1].

This "no-by-default" term is only relevant for the connection, the UI should
call it "no", like openvpn's manual page.

[1] https://git.gnome.org/browse/network-manager-openvpn/commit/?id=62bdd278d69cc396479af9d05a0b776d5ad386bf

BUG:373609