Auto-scan for wireless networks every 15 seconds
NetworkManager plans to reduce periodic scanning in future and
clients are now supposed to scan manually instead. Now the scan
button is removed as it doesn't make sense anymore, because we
now request scan every 15 seconds when either the applet or kcm
is opened. This is implemented the same way in Gnome.

BUG:374552
BUG:372525