Don't assume initialized if there are pending dbus replies
Summary:
ConnectionEditorBase::initialize() does not count the number of pending dbus
replies, so it is currently assuming that if control reaches the end of the
function, we are not waiting for secrets and we can claim the kcm is
initialized. That is not necessarily true, as bug #376739 shows.

If we count the number of replies actually received, we can be sure that there
are no more pending replies and we can then set m_initialized to true at the
end of initialize().

BUG: 376739
FIXED-IN: 5.9.3

Test Plan:
Open Connection KCM from systemsettings5, no longer had the Apply button
enabled. Playing with the KCM still enables the Apply button if something is
changed.

Reviewers: jgrulich

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D4705