Don't duplicate UI option for automatic speed detection
Summary:
Network speed is determined automatically if "auto-negotiate" is set.
NM Spec says if that is not set a speed and duplex should be provided.

The .ui file defaulted to a speed
of 0 with had the special text of "automatic". This isn't actually
automatic, but just an invalid value which
WiredConnectionWidget::isValid (correctly) won't accept

End result is a user can't save any other settings as this widget
reports itself as invalid.

BUG: 383505

Reviewers: #plasma

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D7315