Adjust spacing between text and plotter line in Speed tab of nm applet
Summary:
The text showing the speed of the download in the speed section of the plasma-nm applet was overlapping an sticking to the horizontal lines of the plotter.
the screen shot of the current spacing is as follows:
{F5682035}

I increased the left margin of the graph and also added a line height to the text so that it appears centered with respect to the lines. I asked around the plasma IRC channel and according to @davidedmundson the lines don't correspond to the text, so in my opinion it is safe to make the lines and the text vertically centered to each other.

the screenshot after adding  the spacing is as follows :
{F5682036}

Note: Currently I have only increased the left margin of the graph to accommodated the text assuming the text won't go beyond 4-5 digit and will get converted to MiB from KiB automatically. Ideally this should be dynamic and the graph should auto adjust according to the longest text + margin, So any suggestion for making it that way is welcome.

Reviewers: davidedmundson, #plasma, broulik, jgrulich, ngraham

Reviewed By: jgrulich, ngraham

Subscribers: ngraham, plasma-devel, davidedmundson

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D10183