Use NetworkManager::DeviceStatistics instead of Plasma data engine
Summary:
It's more correct to get information about devices from a single
source, i.e. from NetworkManager. Besides it allows to show the
total number of received/transmitted bytes instead of the current
speed, which is more useful for the user. The current speed can
still be seen on the graph.

This change requires NetworkManagerQt 5.50, because of the fixed
DevicesStatistics::setRefreshRateMs().

Reviewers: #plasma, jgrulich, sitter, jriddell

Reviewed By: jgrulich

Subscribers: broulik, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D14835