Allow the use of  One-Time Password
Summary:
NetworkManager-fortisslvpn allows the user to use a One-Time Password challenge, but this option has not been included in the plasma applet. This change makes allowance for that setting, changing the "Advanced" config dialog to include a checkBox to that effect.

The layout has been taken from the NetworkManager "Advanced" config dialog.

The change has been tested for writing to and reading from the config file, but should still be fully tested

Reviewers: jgrulich

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D17834