Add a popup search bar to the plasma-nm applet
Summary:
If there are not enough connections to fill the scrollview, the search bar is not displayed by default.
If there are, the search bar is shown.
Regardless of that, starting typing will filter the connections list and show the search bar.

I'm uncertain whether not hiding it by default would be a good idea, usability wise.

BUG: 344789
BUG: 394290

Reviewers: #vdg, #plasma, jgrulich, ngraham

Reviewed By: #vdg, jgrulich, ngraham

Subscribers: abetts, jgrulich, ngraham, davidedmundson, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D18607