plasma-nm Connection Icon not showing correct icon when using a bridge
Summary:
This bug has already been reported in bug 397352

Whenever I'm adding a bridge interface to my system, plasma-nm changes the connection icon to a disconnected state even though an active wired or wireless connection is present. In my case, even though I have an active WiFi connection running in the background, as soon as I add the bridge interface, the connection icon changes from 'network-wireless-100' to 'network-wireless-available'.

This is probably happening because when the bridge becomes active, NetworkManager considers it the 'ActivatingConnection' and we are choosing the icon based on the 'ActivatingConnection' without running any checks on it. If the 'ActivatingConnection' returns an invalid object we move onto the 'PrimaryConnection'. I have made this diff to only consider the ActivatingConnection for the icon if the first Device for that connection is a "Wifi", "Ethernet", "Modem" or "Bluetooth". If it's not then move to the PrimaryConnection.

BUG: 397352

Reviewers: jgrulich, #plasma

Reviewed By: jgrulich

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D18689