Remove redundant DNS field
Summary:
BUG: 403546

A DNS field was included on the VPN tab in the initial release to match the
interface of the underlying NetworkManager plugin. This change removes this field and instead
requires the user to use the standard DNS field on the IPv4 and/or IPv6 tabs to enter DNS
servers.
The import function was changed to insert any DNS servers specified in the incoming config
file into the new locations and the export function was changed to output DNS servers from
the IPv4 and IPv6 tabs if they are present but for compatibility with connections created
with the previous version or with the basic NetworkManager plugin, if the IPv4 and IPv6 tabs
do not contain DNS and an (undisplayed) entry on the VPN tab is present it will export from
there.  This allows a user to update a connection to the new format by simply exporting it,
deleting it, and re-importing it.

Reviewers: jgrulich

Reviewed By: jgrulich

Subscribers: ngraham, lbeltrame, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D18867