Change entry of Endpoint UI
Summary:
BUG: 403548

Entry of the Endpoint was originally done in only
one LineEdit which included both the host and port
number. This update changes this so the host (as either
an IPv4, IPv6, or Fully Qualified Domain Name) is
entered in one LineEdit while the port number is
specified in another.

Reviewers: jgrulich

Reviewed By: jgrulich

Subscribers: ngraham, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D19434