No need to set height to implicitHeight
Summary:
There seem to bug a bug in PlasmaComponents.TextField causing to use different
height depending on missing symbol for echo mode when the symbol is used from
a fallback font. While this was fixed in plasma-framework, we seem to be hitting
same issue when we set height to implicitHeight. It shouldn't be necessary to
set the height.

BUG:399155

Reviewers: #plasma, davidedmundson

Reviewed By: #plasma, davidedmundson

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D20436