Add OTP support for openconnect VPN
Summary:
With this patch, One Time Password (OTP) support is added to the openconnect VPN settings and code. It draws the specifications from NetworkManager-openconnect, and presents the same functionality.
The config dialog is modified to include settings for the OTP options
The auth dialog reads and sets the OTP option, updating tokens as needed

Reviewers: jgrulich

Reviewed By: jgrulich

Subscribers: pino, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D18394