Mark "Connection Activated" notifications as low priority
Summary:
In 5.16 onwards, low priority notifications are displayed to the user in a pop-up but
don't get added to the history.

When a network connection is successfully activated, there's nothing further for the
user to have to know or do; it just works and they start using it. As such, there is
no value to keeping the notification in the history; it's just spam. So, this patch
marks that notification as low priority so it doesn't clutter up the history.

Test Plan:
- Apply patch
- Reboot or switch to a different network connection
- See "Connection activated" notification
- Open up notifications history, see nothing there, be happy!

Reviewers: broulik, #plasma, jgrulich

Reviewed By: jgrulich

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D21102