Add the option to configure the realm
Summary: Add an option in the advanced elements dialog to let the user set the connection realm. Realms are authentication user groups that can be defined to give different access rights to the network. The openfortivpn binary and the gnome applets both provide this capability. The diff includes this capability in plasma-nm.

Reviewers: jgrulich

Reviewed By: jgrulich

Subscribers: ngraham, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D21296