Make captive portal notification persistent but low priority
Summary:
The captive portal notification is something that shouldn't time out; when it
becomes visible, it's something the user needs to interact with. I had a
frustrating experience just now trying and failing to click on it before it timed
out.

At the same time, once it's interacted with (either by closing it or clicking the
"Log in" button) there's no need for it to appear in the history, as its history
item is both non-informative and non-interactive.

Accordingly, this patch makes the captive portal notification both persistent and
low urgency (so it doesn't appear in the history).

A similar thing has already been done for Discover in D22429.

Test Plan:
Connected to the in-flight wifi network  on the plane I'm on right now
Captive Portal notification was persistent and didn't show up in the history
after dismissed or logged into via the button on the notification pop-up

Reviewers: #plasma, jgrulich

Reviewed By: jgrulich

Subscribers: apol, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D22467