Exclude non-managed devices from plasma-nm
Summary:
Some network devices are **not** managed by NetworkManager, for example any virtual device created by Docker, any routing devices, etc.

This patch checks if the device is being managed by NM or not before showing it into the applet.

With most certainty, the patch is wrong or the check should be added elsewere, but I thought that would be nice to publish it so we can get the discussion going.

Test Plan: Load the applet with some 10 containers running. The 10 created veth interfaces should not be shown in the applet.

Reviewers: #plasma, jgrulich

Reviewed By: jgrulich

Subscribers: aacid, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D19440