Make "Network Disconnected" notification low priority
Summary:
This is a quintissential low priority notification: nothing bad happens if you miss it,
it's not immediately actionable, and the condition it informs you of is something that
you'll notice anyway. So, let's make it low priority.

Test Plan:
For reasons I do not understand, my tests with this diff did not actually make the
"Network Disconnected" notification stay out of the history as intended. I am unsure
whether this is a local issue on my side or if it doesn't work and needs revision.

Reviewers: jgrulich, broulik

Reviewed By: jgrulich

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D23248