[Applet] Show more information in the Details tab
Summary:
I apologize if I annoy someone with this request, as here https://bugs.kde.org/show_bug.cgi?id=361414 I've seed it has been discussed the issue, and it was done a survey to choice the displayed info.

But, as written in comment #10 of the bug, I don't know why a mac address has more relevance for an average user than an nameserver or a gateway. And usually the average user don't see this section of the plasmoid.

So, I'm submitting to the review process this patch, with hope of being accepted.

It adds just the possibility to see, with the related present info, the current dns (first one) for IPv4 and v6 connectivity, and the current default gateway for the IPv4 connectivity.

FEATURE: 361414
FIXED-IN: 5.18.0

Reviewers: jgrulich

Reviewed By: jgrulich

Subscribers: ngraham, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D24940