[Applet] Fix connecting with a password
Summary:
I could have sworn that I tested this during the initial ExpandableListItem port, but
the code change required to fix it suggests that I either did not, or was smoking crack
at the time. Sorry about the breakage.

Test Plan:
Entering a password in the password field and hit return; password is sent and it tries
to connect

Reviewers: jgrulich, nicolasfella, #plasma

Reviewed By: jgrulich

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D28881