[applet] Make search field always visible
Right now the search field is only visible when specifically activated
by clicking on the search button. This was done because there used to be
a ton of buttons on the toolbar row that made the search field's size
and position look odd when sharing the same row. However with other
recent changes, we no longer have this constraint and there is now
plenty of room for the search field even including some generous
padding to separate it from other controls. So let's put it there.