Make Configuration a proper singleton
Configuration already kind of is a singleton since it is accessed via static methods. Only when used from QML it is instantiated.

By having the methods static we cannot have a changed signal for the properties since a signal can only be emitted on an object.

Make it a singleton for both C++ and QML.