Prevent showing wrong speeds on initial open
Previous attempt was the onCompleted block. However it turns out that
at that point the model still has the old byte amounts.  The order is
showSpeedChanged -> setRefreshRate -> Timer.onCompleted/runningChanged -> model updates
So set the amount to 0 in onCompleted and trigger the Timer on start
(like a Timer with 0 interval). This initial trigger happens after the
 odel gets a new value and we have a correct amount on the first regular
trigger of the Timer.
BUG:420365
FIXED-IN:5.22


(cherry picked from commit d5df9379ef9825a65178d5d2335237cd0f84b90e)