kded: don't show "connection activated" notifications on launch/login
The daemon connects to NetworkManager::ActiveConnection::stateChanged
signals to show notifications, and this signal gets emitted when the
daemon launches to first make its connection. As a result, every time
the user logs in or the plasma-nm kded module is restarted (manually or
due to a crash), a "Connection activated" notification is shown. This is
annoying and redundant since the connection was made with no user action
taken; the user *expects* background processes to work normally and does
not need to be explicitly notified of it.

Plasma-nm already has some logic to suppress these notifications that
would be sent within the first 10 seconds after waking up from sleep,
but it does not handle the case of the same thing happening right when
the daemon is launched (e.g. on login).

This commit makes it handle that use case as well, with the same
approach: suppressing "activated" notifications sent within the first 10
seconds.

BUG: 399973
FIXED-IN: 5.24.2


(cherry picked from commit bd5078d8a5d6dea351339641ab8be3a392039ce8)