Update README.md#Compiling to direct to kdesrc-build
Inspired by my most recent attempt to develop for this specific module. I used the build instructions and ended up soft-bricking my desktop. This was resolved with a little debugging and foresight, however it may be better to point to the KDE developer kdesrc-build best practices guide.

The bug in question was using `sudo make install` (final step in README.md#Compiling), which installed:

`/usr/lib/qt/plugins/plasma/kcms/systemsettings_qwidgets/kcm_networkmanagement.so`

on Manjaro Linux. However this is not the correct path and caused `systemsettings5` to not recognize or register/load the `kcm_networkmanagement.so` UI or list the module in the CLI. The solution was to delete this file and then reinstall `plasma-nm`.