applet: port away from contextMenu and use contextualActionsModel
With changes to ExpandableListItem, we can now define both contextual
actions and a custom content view. This means we can put the actions
that are currently in the context menu in the contextual actions list
instead, so they will be displayed in the expanded view above the custom
content.

This yields several benefits:
- Makes these items more consistent with other expandable list items in
  other widgets
- Removes the use of a context menu, which goes against the purpose of
  this component (to be fully usable with only left-clicks or touches)
- Removes a PlasmaComponents2 import