Default to creating system-wide connections if allowed
The IWD backend for NetworkManager does not support user-owned
connections and will refuse to connect when trying to create such a
connection. Since it is impossible to determine that the IWD backend is
being used, instead just always default to creating system-wide
connections if we are allowed to. This matches the behaviour of Gnome's
network-manager-applet where the same change was made (for the same
reason).

For reference, see
https://gitlab.gnome.org/GNOME/network-manager-applet/-/merge_requests/98
,
https://gitlab.freedesktop.org/NetworkManager/NetworkManager/-/issues/743
and
https://gitlab.freedesktop.org/NetworkManager/NetworkManager/-/issues/845#note_1311483