Add UiUtils::setConnectionDefaultPermissions
This sets up default permissions correctly for a connection. When either
configuration indicates we should prefer system connections, KWallet
isn't enabled or we're running a live image we'll try to create a system
connection. However, if we're not allowed to create system connections,
then fall back to creating user connections.

Additionally, this sets the default password storage to store passwords
system-wide if we're creating system-wide connections, since storing
those passwords locally leads to awkward behaviour where a system
connection can't really be reused since the password is not available.