Show current MAC address instead of permanent MAC address in details view
Current behavior is for the details view to show the permanent
MAC address of the card. This currently creates a missunderstanding
in which the user might think that the MAC address in use is, in fact,
the one shown there, while this might not be the case if the
MAC addres has been cloned/changed/spoofed. This leads the user to
believe that the change to his MAC address did not work or to not be
able to recognize his own traffic if he expect to see that MAC on
sniffed traffic for example, and many other scenarios.

NOTE: the current fix is applied only to the wireless device as it I
am unable to test the wired device code block and would refrain from
committing untested code, but it is highly advisable to add those too.

BUG: 456119
(cherry picked from commit 7908728eb34c4311b2a2ae9b17e5cdba9f587c18)