applet: Fix enabled states of Wi-Fi and WWAN checkboxes in Edit Mode
We don't have full control over `enabled` property all the time, as it
inherits its state through parent-child chain in QML. Instead of
reusing it for custom logic, just introduce a new property for that.

This patch replaces assignments with bindings, and drops two extra
signal handlers which are not needed anymore because the property they
were updating is now a reliable readonly boolean.