applet: Use deliberate resize duration for expandable list items
These items now animate in a slightly different way in Plasma 6, and
because of their large height when expanded, it's nicer to make the
animtion a bit faster so the user isn't waiting so long when an item is
expanded.

Move duration is unchanged because we don't want animations when moving
between items.