Prepare & adapt includes to no <QtDBus> include being used in deps
QStringBuilder is being explicitly included to match the old state where
<QtDBus> also pulls in <QtCore> which again pulls in <QStringBuilder>
without condition, other than <QString> which only does when
either QT_USE_FAST_OPERATOR_PLUS or QT_USE_QSTRINGBUILDER is set, which
is not the case currently for plasma-nm.
The use of operator% for string concatenation in the respective files
though needs <QStringBuilder> being included, so it has to be done
explicitly now.

GIT_SILENT