Show captive portal login page in a dedicated window
This streamlines the entire process and separates it more
clearly from whatever else might be going on on the users
machine and browser. It also gives us the opportunity to
give us more context to what is happening (for example
with the window title).
A more niche case are some devices (such as the steamdeck)
that  might come without having a browser preinstalled leaving
the user in an impossible situation requiring a browser to gain
network access to be able to install a browser.
If for whatever reason the authentication page might not work
in QWebEngine the current page can still be opened in a browser
via a button.