Port to upstream QML components where possible
Basically just PlasmaComponents3.Label -> QQC2.Label because everything
else has already been ported.

We can do this because this component only existed in the past to use
special Plasma colors, but no visual styling. And the color stuff is
obsolete because Kirigami colors work correctly in Plasma now.

Everything was tested with a dark Plasma theme and a light color scheme
to ensure no regressions.