applet: Untangle and improve layout of the network speed graph
Instead of reserving leftMargin on the graph for axis labels and then
anchoring to the graph's right side to fit labels in that gap, lay it
out in a more straightforward way: axis is anchored to parent.left,
bind simple expression for width, and anchor the graph with a small
fixed margin. Also, top-level Layout.leftMargin makes the whole thing
not stick to the delegate's edges, and is now consistent with the
Legend item below.