Port to KUrlRequester::setNameFilters
Unlike !279 nameFilters does exist in KUrlRequester, and the filter does work:

![image](/uploads/88a3d0419a413a8d458f6a26bd35b663/image.png)

~~Marked as draft since I have not requested approval for a string freeze exception yet just in case anyone has objections to the wording, in particular the two fields in the OpenVPN one that says "Local peer's private key in .pem format." confuses me since it uses the same set of file extensions as the "Certificate authority (CA) file in .pem format." field, can't tell if it's intentional or not.~~

~~Ideally we'd just have the filter show the extensions with no wording like how it was with KF5, but just using `<string>*.der *.pem *.crt *.cer</string>` or `<string>(*.der *.pem *.crt *.cer)</string>` unfortunately does not work, seems like that's actually a bug as per the discussion in https://invent.kde.org/frameworks/kio/-/merge_requests/1503 but if so I'm not sure whereabouts to report that.~~

With https://invent.kde.org/plasma/plasma-integration/-/merge_requests/128 nameless filters work again, so with that a string freeze approval is no longer required. Ideally then this MR should be merged after that one, but the worst that merging this first will do is to break the filter, but selecting files still works, and the compilation problem is still fixed either way.