Add Remote Identity, password options and password save/load
All changes applies to the StrongSwan plugin.

- Add a field for Remote Identity. This feature exists in
nm_applet (GTK) from StrongSwan. It is needed for certain
servers that use certificates from internal CA, otherwise
the authentication may fail.

- Add password options. With this, frequently changing passwords
policies can be easily configured with "don't save/ask always"
option. Other plasma-nm plugins already have this, as well the
GTK nm_applet.

- Add password save/load. Sometimes, when you provide a password,
it is stored somewhere and it isn't easy to remove it and force
the interface to ask it again. With options enabled, this isn't a
must, but it was a bit strange if left this issue alone. Now it
saves and loads the password in the way charon-nm understands.