Add WebAuth support for SAML-based authentication
In some deployments, users are required to use hardware security devices
to authenticate themselves e.g. YubiKey.

Since the 6.7.0 release, QtWebEngine has added support for WebAuth
handling. When a page requests WebAuth authentication, the request is
passed to the host application through a signal. The application must
then handle the request by, for example, displaying a dialog to the
user.

The QtWebEngine project implements a reference host application:
simplebrowser, which implements such a dialog. This patch imports this
code, makes various style improvements and connects it to the QWebView.

Signed-off-by: Joel Holdsworth <joel@airwebreathe.org.uk>