Add WireGuard config field for interface name
Fixes ~~BUG: 412795~~ BUG: 479179

Previously the WireGuard configuration took the connection name as the
interface name. This had a host of problems, especially that the default
connection name was not valid for an interface name, and that an invalid interface
name would result in losing all of your changes.

This commit adds a required "Interface name" field (with proper validation). This behavior mirrors `nm-connection-editor`.

![New connection screenshot](/uploads/f4dbd88a73d207664ff1e096b30374e3/image.png)

BUG: 479179