kcm: Use sentence case in json description
The description field in the json file tor the kcm currently uses mixed
capitalization: "Edit your Network Connections". This is neither of the
generally recommended sentence or title cases. In particular, the
recommended case for subtitles or tooltips (the most likely use of this
description) is sentence case, which the majority of kcms seem to use.
The word "your" also seems both superfluous and somewhat incorrect, as
system-wide changes (i.e. for other users as well, not just you) can
also be made here.

This change makes the json file use sentence case and omits "your":
"Edit your network connections"