applet: remove now-unneeded custom CompactRepresentation
Part of https://invent.kde.org/plasma/plasma-workspace/-/issues/85

It only existed because the standard CompactRepresentation had no
functionality to handle middle-clicks. But it's gained that now as of
https://invent.kde.org/plasma/plasma-desktop/-/merge_requests/2473
and https://invent.kde.org/plasma/libplasma/-/merge_requests/1200,
so let's make use of it and delete some unnecessary custom code!