802.1x: if scheme is not 'file' leave chars in percent-encoded form
For example, if `private-key`, `ca-cert` or `client-cert` scheme is `pkcs11`, according to RFC 7512 and RFC 3986,
a space char and characters outside the US-ASCII must be encoded
in octets according to the UTF-8 character encoding and then percent-encoded.

Otherwise, without this patch, if you change and save something in `plasma-nm`, then `%20` will be replaced with a space in the `nmconnection` file with the connection, which will make the URI incorrect and break the connection.

These parameters (`private-key`, `ca-cert`, `client-cert`) could has not only `file:///` scheme according to [NM docs](https://networkmanager.pages.freedesktop.org/NetworkManager/NetworkManager/settings-802-1x.html).

<details><summary>Example</summary>
![image](/uploads/62739da6f03149ccc1b5c45e3a45105c/изображение.png){width=506 height=423}
</details>