applet/ConnectionListPage: disable unused scrollbar
The connection list has the ScrollBar at its default AsNeed policy,
but the content is resizing so that a horizontal scrollbar is never
needed. This is not only some extra unnecessary work, but also causes a
binding loop when the popup is closed.

Instead, just mark the horizontal ScrollBar as AlwaysOff and avoid the
extra work and binding loop.