kcms/kcm_connections: port to ItemDelegate
The connection list is not keyboard and screen-reader accessible. This
is difficult to fix at present, as it uses a custom delegate that does
not follow our usual ways of implementing this, and also causes various
visual inconsistencies such as a different hover feedback from other
kcms.

The custom delegate doesn't seem to do anything we could not do with
ItemDelegate and IconTitleSubtitle, so we port to that. There should not
be functionality or major visual changes, except for some details
becoming more consistent with ListViews elsewhere. We also port the
AddConnectionDialog so we can remove the base component for the old
delegate.