kcms/kcm_connections: improve keyboard and screen reader a11y
The main connection list has some keyboard navigation shortcomings,
including the absence of visual feedback that makes keyboard-based use
very opaque, and for some interface elements (such as icon-only buttons)
the correct Accessible names and/or roles are not set.

This change makes keyboard navigation of the list (and whole sidebar)
work mostly as expected, and exposes the the correct roles and/or
existing strings to the Accessibility framework where necessary.

BUG: 489947