libs: Fix hotspot names changing not syncing between instances of Handler
Fix https://invent.kde.org/plasma/plasma-mobile/-/issues/462

I'm not sure this MR is clean but it fix the bug.
When we have multiple Handler as plasma mobile (One for settings and one for quicksettings), the changes on KSharedConfig is not synchronized.
So when we change the name of hotspot on settings and then, we enable hotspot from quicksettings, it used the old hotspot name.

I need to add Notify flags and grp.sync() to send a notification in dbus and from my test. KConfigWatcher need to be created as instance variable (static doesn't work).

I see also, it not need signal to be attached. I think, it's because `static KSharedConfig::Ptr openConfig` return a static reference and use the same reference as writer methods.

**Note**:
 I see the mutex is not used on all methods which write in KSharedConfig and not sync group too. I don't know if is a mistake because i still learning QT and KDE stack.
 if is a mistake, i can add sync and lock mutex in all methods.

**Note 2**:
 if my solution is correct, it can be good to add Flags::Notify on all methods which write into KSharedConfig