applet: Fix list interactions with the password field open
Clicking connect or disconnect on a connection item would not give any
feedback if a password was being entered in another connection item
since model updates were disabled.

This commit changes the behavior so that clicking connect/disconnect 
closes other open password fields, ensuring that either one password
field is open, or the model begins updating if we can freely connect or
disconnect the other item.

BUG: 477655
FIXED-IN: 6.5.0