feat: make wifi connections global by default
This change simply alters how a connection is created so that it doesn't
specify that it is restricted to a user, and will therefore allow the
connection to activate for any user -- including before login.

Benefits:

- Wifi can be active before login for enterprise / government use cases
- Connection doesn't have to be made for each user on a shared machine
- KISS can set up the wifi during the out of box experience

The connection config is still stored at
`/etc/NetworkManager/system-connections` and still requires root access
to view or modify.

Obsoletes https://invent.kde.org/plasma/kiss/-/merge_requests/43
Resolves https://invent.kde.org/kde-linux/kde-linux/-/issues/236
Resolves https://invent.kde.org/plasma/kiss/-/issues/13