OpenConnect KCM: fix saving/loading of empty reported OS
If no reported OS (i.e. system default) was selected, an empty string
would be inserted into the settings map. However, the loading code only
checked for presence of the key in the map, not whether it's empty, with
the result that the combobox would show the last entry (iOS) instead of
the default blank one. This fixes both the loading part (it now checks
for empty strings and also only selects the last entry if the value
matches instead of unconditionally), and the saving part (the setting is
only inserted into the map if it's not empty)